+++ crates/lkjscript-core/src/outcome/structural.rs
@@ -2,8 +2,8 @@
 /// These observations do not grant or deny language validity.
 #[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
 pub struct StructuralSnapshotMetrics {
-    pub nodes: u32,
-    pub fields: u32,
+    pub nodes: u64,
+    pub fields: u64,
     pub aggregate_bytes: u64,
     pub string_bytes: u64,
     pub path_bytes: u64,
