+++ crates/lkjscript-core/src/structural/value_runtime/events.rs
@@ -32,7 +32,7 @@ pub enum StructuralEventKind {
 pub struct StructuralEvent {
     pub sequence: u64,
     pub kind: StructuralEventKind,
-    pub subject: u32,
+    pub subject: u64,
     pub amount: u64,
 }
 
@@ -79,7 +79,7 @@ impl StructuralEventLog {
 
     /// Retains diagnostics opportunistically. Allocation failure must not stop
     /// ownership cleanup or alter structural semantics.
-    pub(super) fn record(&mut self, kind: StructuralEventKind, subject: u32, amount: u64) -> bool {
+    pub(super) fn record(&mut self, kind: StructuralEventKind, subject: u64, amount: u64) -> bool {
         if self.records.try_reserve(1).is_err() {
             self.omitted = self.omitted.saturating_add(1);
             return true;
@@ -107,18 +107,18 @@ pub struct StructuralValueRuntimeMetrics {
     pub releases: u64,
     pub stale_rejections: u64,
     pub object_slots_reused: u64,
-    pub live_objects: u32,
-    pub peak_live_objects: u32,
+    pub live_objects: u64,
+    pub peak_live_objects: u64,
     pub destinations_created: u64,
     pub destinations_completed: u64,
     pub destinations_aborted: u64,
     pub destination_fields_initialized: u64,
     pub destination_cleanup_work: u64,
-    pub live_destinations: u32,
+    pub live_destinations: u64,
     pub views_created: u64,
     pub views_ended: u64,
-    pub live_views: u32,
-    pub peak_live_views: u32,
+    pub live_views: u64,
+    pub peak_live_views: u64,
     pub string_bytes_allocated: u64,
     pub string_bytes_live: u64,
     pub string_bytes_cloned: u64,
@@ -131,13 +131,13 @@ pub struct StructuralValueRuntimeMetrics {
     pub payload_bytes_peak: u64,
     pub clone_nodes: u64,
     pub release_work: u64,
-    pub release_backlog: u32,
+    pub release_backlog: u64,
     pub sealed_publications: u64,
     pub zero_copy_adoptions: u64,
     pub sealed_acquisitions: u64,
     pub sealed_releases: u64,
-    pub live_sealed_domains: u32,
-    pub live_sealed_owners: u32,
+    pub live_sealed_domains: u64,
+    pub live_sealed_owners: u64,
     pub sealed_release_work: u64,
     pub sealed_nodes_reclaimed: u64,
     pub copied_publication_bytes: u64,
